[refactor] #181 self-invocation 제거 - deleteExpiredRequest를 RequestExpiryService로 분리
